+++ libfft.h
@@ -1,7 +1,8 @@
 #ifndef LIBFFT_H
 #define LIBFFT_H
 #include <complex.h>
+#include <stdint.h>
 
-void fft(float complex *data, const unsigned int N);
+void fft(float complex *data, const uint16_t N);
 
 #endif
