+++ main.c
@@ -3,18 +3,18 @@
 #include "libfft.h"
 #include "libsiggen.h"
 
-static void __write_complex_to_file(const float complex *buffer, const unsigned int N, const char *filename)
+static void __write_complex_to_file(const float complex *buffer, const uint16_t N, const char *filename)
 {
     FILE *fp = fopen(filename, "w");
     fprintf(fp, "re,im\n");
-    for (unsigned int i = 0; i < N; i++)
+    for (uint16_t i = 0; i < N; i++)
     {
         fprintf(fp, "%f, %f\n", crealf(buffer[i]), cimagf(buffer[i]));
     }
     fclose(fp);
 }
 
-static void __run_fft(float complex *buffer, const unsigned int N, const char *name)
+static void __run_fft(float complex *buffer, const uint16_t N, const char *name)
 {
     char filenamebuffer[50];
     sprintf(filenamebuffer, "%s_original.csv", name);
@@ -28,7 +28,7 @@ static void __run_fft(float complex *buffer, const unsigned int N, const char *n
 
 int main()
 {
-    const unsigned int MAX_N = 1024;
+    const uint16_t MAX_N = 1024;
     float complex buffer[MAX_N];
 
     generate_saw(buffer, MAX_N);
